+++ CHANGELOG.md
@@ -9,12 +9,16 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 ---------------------------------------------------------------------------
-## 1.2.5 - 2022-11-22
+## 1.2.5 - 2022-11-??
 
-Incremental improvements and new preview features.
+Incremental improvements, new preview features and bug fixes.
 
 ### New features:
 - Add attribute returned for getSystem: *isPublic*.
+- Check authorization and validate execSystemId and archiveSystemId at application creation time.
+
+### Bug fixes:
+- Fix bug in setting of sharedAppCtx for getApp.
 
 ---------------------------------------------------------------------------
 ## 1.2.4 - 2022-11-14